[rename]: CategoryResult 페이지 레이아웃 분리
