+++ src/layouts/CategoryResult/CategoryList/index.module.scss
